Network: Add Option to set custom DNS server (part 1)
Add an option in the Network & Internet Settings fragment to
allow setting a custom DNS server (IPv4 and IPv6), which will
override (via iptables port 53 NAT) the existing ISP or WiFi
provided DNS. It uses the defined properties of the 'z_iptables'
service.

This part creates the 'body': A new preference and an own
preference screen fragment, but yet without the ability to
change input and verify the IPv4/IPv6 values, but with a basic
preference controller to show the IPv4 settings to allow some
initial testing.

The option shall only be available for the device owner (i.e.
not for secondary users) and all settings are to be reset, when
the option "Reset network settings" is executed.

For TESTING-ONLY purposes (TO BE REMOVED LATER !!), Cloudflare's
DNS 1.1.1.1 / 2606:4700:4700::1111 are being used as hard-
coded constants.

Change-Id: I85c12a10b8320ac1a2204de8e3c95f8096f6c296